feat: setup basic parse functions in helper file
