+++ source-code/website/src/pages/editor/@host/@owner/@repository/index.page.tsx
@@ -41,6 +41,10 @@ function TheActualPage() {
 	 *    }
 	 *  }
 	 */
+
+	createEffect(() => {
+		console.log(resources)
+	})
 	const messages = createMemo(() => {
 		const result: {
 			[id: string]: {
